+++ apps/web/package.json
@@ -56,6 +56,7 @@
     "date-fns": "^3.3.1",
     "dayjs": "^1.11.7",
     "decimal.js": "^10.4.3",
+    "file-type": "^19.0.0",
     "lucide-react": "^0.259.0",
     "next": "^14.2.0-canary.58",
     "react": "^18.2.0",
